# chunkmatches: reuse last calculated column when filling