--- a/contentprovider_test.go
+++ b/contentprovider_test.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"fmt"
 	"testing"
+	"testing/quick"
+	"unicode/utf8"
 
 	"github.com/google/go-cmp/cmp"
 )
@@ -325,5 +327,83 @@ func TestChunkMatches(t *testing.T) {
 				t.Fatal(diff)
 			}
 		})
+	}
+}
+
+func BenchmarkColumnHelper(b *testing.B) {
+	// We simulate looking up columns of evenly spaced matches
+	const matches = 10_000
+	const match = "match"
+	const space = "         "
+	const dist = uint32(len(match) + len(space))
+	data := bytes.Repeat([]byte(match+space), matches)
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		columnHelper := columnHelper{data: data}
+
+		lineOffset := 0
+		offset := uint32(0)
+		for offset < uint32(len(data)) {
+			col := columnHelper.get(lineOffset, offset)
+			if col != offset+1 {
+				b.Fatal("column is not offset even though data is ASCII")
+			}
+			offset += dist
+		}
+	}
+}
+
+func TestColumnHelper(t *testing.T) {
+	f := func(line0, line1 string) bool {
+		data := []byte(line0 + line1)
+		lineOffset := len(line0)
+
+		columnHelper := columnHelper{data: data}
+
+		// We check every second rune returns the correct answer
+		offset := lineOffset
+		column := 1
+		for offset < len(data) {
+			if column%2 == 0 {
+				got := columnHelper.get(lineOffset, uint32(offset))
+				if got != uint32(column) {
+					return false
+				}
+			}
+			_, size := utf8.DecodeRune(data[offset:])
+			offset += size
+			column++
+		}
+
+		return true
+	}
+
+	if err := quick.Check(f, nil); err != nil {
+		t.Fatal(err)
+	}
+
+	// Corner cases
+
+	// empty data, shouldn't happen but just in case it slips through
+	ch := columnHelper{data: nil}
+	if got := ch.get(0, 0); got != 1 {
+		t.Fatal("empty data didn't return 1", got)
+	}
+
+	// Repeating a call to get should return the same value
+	// empty data, shouldn't happen but just in case it slips through
+	ch = columnHelper{data: []byte("hello\nworld")}
+	if got := ch.get(6, 8); got != 3 {
+		t.Fatal("unexpected value for third column on second line", got)
+	}
+	if got := ch.get(6, 8); got != 3 {
+		t.Fatal("unexpected value for repeated call for third column on second line", got)
+	}
+
+	// Now make sure if we go backwards we do not incorrectly use the cache
+	if got := ch.get(6, 6); got != 1 {
+		t.Fatal("unexpected value for backwards call for first column on second line", got)
 	}
 }