@keegancsmith 2024-01-09
Here are my journal notes for the pprof part of this investigation. It contains useful information for future performance debugging.

<span class="timestamp-wrapper"><span class="timestamp">[2024-01-09 Tue 10:51] </span></span> Gonna try grab a pprof during the search. It is only 2 seconds, so unsure how useful it will be. Maybe can try adding a loop.

I can just visit <https://sourcegraph.sourcegraph.com/-/debug/proxies/indexed-search-0/debug/pprof/profile>

I then used devtools to `Copy as cURL` the request. I then did the same for making the search request. For both I copied into a bash function and added the `--silent --show-error` flags for my sanity. The script looked like this, curl flags removed.

```
#!/usr/bin/env bash

set -e

function fetch_profile {
    echo "profile start"
    curl 'https://sourcegraph.sourcegraph.com/-/debug/proxies/indexed-search-0/debug/pprof/profile' > /tmp/cpu.pprof
    echo "profile done"
}

function search {
    echo "search start"
    curl 'https://sourcegraph.sourcegraph.com/search/stream?q=context%3Aglobal%20repo%3Agithub.com%2Fsourcegraph%2Fsourcegraph%20%20content%3A%22dev%22%20&v=V3&t=newStandardRC1&sm=0&display=1500&cm=t&trace=1&feat=search-debug' > /dev/null
    echo "search done"
}

fetch_profile &

while jobs %%; do
    search
done
```

Next thing I needed was the zoekt binary used. I did this by getting the version of s2 at <https://sourcegraph.sourcegraph.com/__version> and then getting the binary from the docker container:

```
docker pull sourcegraph/indexed-searcher:257084_2024-01-09_5.2-9efa6c7e2efb
docker create sourcegraph/indexed-searcher:257084_2024-01-09_5.2-9efa6c7e2efb
docker cp 74ad840ae6a8f51ddec7ff4a382660ca8a62bb4d47d7730f355d71f9b68cde15:/usr/local/bin/zoekt-webserver /tmp/
docker rm -v 74ad840ae6a8f51ddec7ff4a382660ca8a62bb4d47d7730f355d71f9b68cde15
```

    go tool pprof -http 127.0.0.1:6062 zoekt-webserver cpu.pprof

Then so I could use source code listing

    go tool pprof -trim_path external/com_github_sourcegraph_zoekt/ -source_path ~/src/github.com/sourcegraph/zoekt zoekt-webserver cpu.pprof

Turns out we spend 96.50% in `utf8.RuneCount` inside of `fillContentChunkMatches`. This is an issue only with chunk matches, not the original format. We calculate the column, and it has a hidden `O(n^2)` algorithm in it! If you have a long line, we basically do a `O(n)` operation on that line per match, where `n` is the line length.

<https://github.com/sourcegraph/zoekt/blob/e92f6c764b567d9b359633fa4045cfba9d37eb3c/contentprovider.go#L309>

```
(pprof) list fillContentChunkMatches
Total: 25.15s
ROUTINE ======================== github.com/sourcegraph/zoekt.(*contentProvider).fillContentChunkMatches in contentprovider.go
      10ms     24.36s (flat, cum) 96.86% of Total
         .          .    291:func (p *contentProvider) fillContentChunkMatches(ms []*candidateMatch, numContextLines int) []ChunkMatch {
         .       20ms    292:	newlines := p.newlines()
         .       30ms    293:	chunks := chunkCandidates(ms, newlines, numContextLines)
         .          .    294:	data := p.data(false)
         .          .    295:	chunkMatches := make([]ChunkMatch, 0, len(chunks))
         .          .    296:	for _, chunk := range chunks {
         .       20ms    297:		ranges := make([]Range, 0, len(chunk.candidates))
         .          .    298:		var symbolInfo []*Symbol
         .          .    299:		for i, cm := range chunk.candidates {
         .          .    300:			startOffset := cm.byteOffset
         .          .    301:			endOffset := cm.byteOffset + cm.byteMatchSz
      10ms       20ms    302:			startLine, startLineOffset, _ := newlines.atOffset(startOffset)
         .          .    303:			endLine, endLineOffset, _ := newlines.atOffset(endOffset)
         .          .    304:
         .          .    305:			ranges = append(ranges, Range{
         .          .    306:				Start: Location{
         .          .    307:					ByteOffset: startOffset,
         .          .    308:					LineNumber: uint32(startLine),
         .     12.42s    309:					Column:     uint32(utf8.RuneCount(data[startLineOffset:startOffset]) + 1),
         .          .    310:				},
         .          .    311:				End: Location{
         .          .    312:					ByteOffset: endOffset,
         .          .    313:					LineNumber: uint32(endLine),
         .     11.85s    314:					Column:     uint32(utf8.RuneCount(data[endLineOffset:endOffset]) + 1),
         .          .    315:				},
         .          .    316:			})
         .          .    317:
         .          .    318:			if cm.symbol {
         .          .    319:				if symbolInfo == nil {
```