@jtibshirani 2024-01-09
> I think that equivalent logic exists in searcher. We should probably implement this same thing there as well.

Assigning this to myself! Said differently: please don't work on this concurrently, as it will cause a lot of conflicts with my in-progress searcher optimizations (https://github.com/sourcegraph/sourcegraph/issues/59038) :)